rinad: ipcm: catalog: move to a separate compilation unit
ipcm compilation unit is already big enough, so that it makes
sense to move the catalog into a separate one.
