+++ rinad/src/ipcm/Makefile.am
@@ -45,7 +45,8 @@ ipcm_SOURCES  =								\
 	configuration.cc		configuration.h			\
 	flow-alloc-handlers.cc		flow-alloc-handlers.h		\
 	policies-handlers.cc						\
-	dif-template-manager.cc		dif-template-manager.h
+	dif-template-manager.cc		dif-template-manager.h		\
+	catalog.cc			catalog.h
 
 test_empty_SOURCES  =				\
 	test-empty.cc
